chore(connect-explorer): remove all remaining circular imports
